auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0